Check stderr of ngspice
Previously the stderr would just fly by as all the tests were
being executed. Now it's actually captured, and checked. There
are a couple of harmless warning which can be safely ignored,
but anything else results in an error. That way if there are any
issues ngspice detects, but for whatever reason doesn't cause it
to stop the sim or exit early, then the test will still fail.