Improve eqn_solver.py performance
Rather than using a dict and then build the matrix with lil_matrix
entry by entry, instead keep track of each (row, col, data) coefficient
addition, and then build a coo_matrix from that, and then convert
it to a csc matrix. For some reason directly building the csc matrix
takes longer. This offers a significant performance boost to the
simulation. override_variable is removed, as it just complicates matters
and it turns out it wasn't ever needed.