Add an addition program and related functionality
This took a fair bit of effort to get the required functionality in the
assembler. Discovered a few bugs along the way, such as JUMPs always
pointing to an instruction past their destination due to the increment
occuring after setting PC. In the actual CPU that may be the actual
functionality, I'll have to decide whether I want to expend effort or
making logic to stop that from happening. Also 'add' an addition
program... geddit?