Add addition to the common asm library
This required a few tweaks to the assembler, including adding a
read_var() method for indirection over the various sections, and fixing
a bug where the context of a code value would be determined by where the
command is defined rather than where the value originated. This SECTION
stuff should make it easier to build a good library for more complex
programs.