Allow program debugger to recognise labels
Both in address decoding and for the current instruction. Both local and
global. To facilitate local labels, information about the current global
label for the given context is bundled into ProgramTraceback, which is
starting to become a pretty heavy object.

Also fix a bug with IF which would run the code value with the context of
the command rather that of the value. This then caused problems with the
'needed' system which meant I modified the UP command without knowing
precisely what was going on. Using the debugger actually helped me
figure it out with the tracebacks, even though the issue was compile
side. Both the IF and UP commands are now fixed.