Add INTERNAL_COMMAND define type
This is used by the debugger to display the 'current line' as what the
program write has written rather than always showing the implementation
of INSTRUCTION_DATA. There's no general method to do this a priori as
what separates internal library code from the rest of the library and
the program is subjective. The current line also influences the context
used for e.g. local label resolution.

While I was at it, I also renamed some of the commands to be more
consistent. Until I write a language server extension for VS Code I'll
have to lean heavily on memory was the library grows larger, so having
consistent naming is very important.