Implement binary comparison operations
The design I settled on was one binary table which encodes the 3 states
of less than, equal to, and greater than for each pair of words. Then a
second instrucion must be used to extract the result of the desired
operation (less than, less than or equal, equal, not equal, greater than
or equal, greater than). While this does require two instructions rather
than one (as in a binary table for each operation), the initialisation
time and memory footprint is definitely reduced.