Implement prime number generator
This is the final milestone for this phase of the project. The
efficiency and complexity are horrrible, but it'll do as a proof of
concept. The only thing I think I'll look at is maybe changing the
architecture to make CALL more efficient.

I implemented the big_int_10_cmp function for this, which turned out
better than I expected.